+++ app/assets/scss/mobile.scss
@@ -1,0 +1,1 @@
+// placeholder para estilos mobile
